Extract session logic for remote attestion, GH issue ref #2741
