+++ grpc_unary_attestation/src/client.rs
@@ -14,14 +14,12 @@
 // limitations under the License.
 //
 
-use crate::{
-    proto::{unary_session_client::UnarySessionClient, UnaryRequest},
-    SessionId,
-};
+use crate::proto::{unary_session_client::UnarySessionClient, UnaryRequest};
 use anyhow::Context;
 use oak_functions_abi;
-use oak_remote_attestation::handshaker::{
-    AttestationBehavior, ClientHandshaker, Encryptor, ServerIdentityVerifier,
+use oak_remote_attestation::{
+    handshaker::{AttestationBehavior, ClientHandshaker, Encryptor, ServerIdentityVerifier},
+    sessions::SessionId,
 };
 use tonic::transport::Channel;
