+++ grpc_unary_attestation/src/lib.rs
@@ -16,8 +16,6 @@
 
 //! Remote Attestation handshake protocol implemened using bidirectional streaming gRPC.
 
-pub type SessionId = [u8; 8];
-
 pub mod proto {
     #![allow(clippy::return_self_not_must_use)]
     tonic::include_proto!("oak.session.unary.v1");
