+++ remote_attestation/rust/Cargo.toml
@@ -15,6 +15,8 @@ bytes = { version = "*", default-features = false }
 log = "*"
 prost = { version = "*", default-features = false, features = ["prost-derive"] }
 ring = { path = "../../third_party/ring", default-features = false }
+lru = "*"
+oak_utils = { path = "../../oak_utils"}
 
 [build-dependencies]
 prost-build = "*"
