+++ remote_attestation/rust/src/lib.rs
@@ -27,5 +27,6 @@ pub mod crypto;
 pub mod handshaker;
 pub mod message;
 pub mod report;
+pub mod sessions;
 #[cfg(test)]
 mod tests;
